Adding optimization to skip "doubling back" routes
If the route loops back on itself without picking
up or dropping off an alien, then that route is
topologically equivalent to a shorter route where
it goes directly from the start of the loop-back
to the end of it. Therefore, we can consider any
route that loops back on itself to be a "dead end",
because we'll eventually test that shorter version,
or already have.

Adding this optimization saves a *lot* of time on
levels with lots of empty space, where the solver
otherwise wastes a lot of time testing out every
possible way to fill up the empty space.

For now the loopback test is based on whether
we become adjacent to an earlier step in the route.
It might be possible to squeeze out some more
performance with a smarter version that takes into
account junctions and/or warps.