Earlier detection of when a vital cell is unreachable
The code already checked whether any vital cell became
completely blocked off. This code additionally checks
whether the train would be forced to dead-end when
approaching the vital cell.

Example:

 ###.
 #O..
 ###.

^^ In this scenario, for the train to reach the O it would
have to travel into a square with no exits. So we can safely
say that if the map looks like this, it's unsolvable.

This makes little difference on some maps, and lots of difference
on others. Particularly maps with many aliens and houses next
to each other and/or next to walls. (Like Nova 7!)

This approach still has its limitations; it's basically a simple
special case of a "bottleneck" in the map. However, there are
more complex cases, such as if there is plenty of space around
the vital spot, but a bottleneck elsewhere:

 #####.
 #...#.
 #O....
 #####.

However, these are harder to identify algorithmically.