WIP using Suurballe's algorithm to check for solvability
See: https://en.wikipedia.org/wiki/Suurballe%27s_algorithm

The existing "region check" code that checks each step for solvability,
and the recent extensions that checks for "space to turn around in",
were really pointing at the question of whether it's possible to draw
a path from the front car, to the vital space in question, to the exit.
If there's not, then you know the current partial solution cannot be
correct.

I did some digging into this to see if there were any known algorithms
for this problem. It turns out that it can be modelled as the problem
of finding "node disjoint paths" from A to B. ("Node disjoint" means
paths that do not share any nodes in common.) You just add an imaginary
"start" node that connects to the first car, and the exit. Then you
look for two disjoint paths from that "start" node to the vital node.

And, it turns out there is an existing well-documented algorithm for
this problem, Suurballe's algorithm. So, I'm going to try to implement
that, using the graphology library to handle the graph manipulations.

Unfortunately the version of Suurballe's algorithm that's best
documented, is the version that finds edge-disjoint paths in a
directed graph. What I want is node-disjoint paths in an undirected
graph.

But it turns out there, too, you can transform the graph to match.
To make it into a directed graph, you just replace each undirected
edge with two directed edges going opposite directions. (This actually
works better for me anyway, because then I can put just one outgoing
directed edge on the non-navigable vital cells, like aliens).

And to turn the edge-disjoint path finding algorithm into a node-disjoint
path finding algorithm, you split each node into two: an "input" node,
and an "output" node. You map all the original node's incoming edges to
the "input" node, and all its outgoing edges to the "output" node, then
you draw a single directed "connector" edge from the input node to the
output node. Then you run Suurballe's algorithm and if the path goes
through one of those "connector" edges, it's the same as if it went through
the original node.