Using a faster version of Surballe's algorithm (still too slow)
I've made some tweaks to the version of Surballe's I'm using,
but it's still quite a bit slower than not using it.

1. Changed from vertex-disjoint version to edge-disjoint
    version. This means I don't have to split each node of
    the graph in two, so it cuts the number of edges in half.
2. Removed the code that reweighted every plot in the graph.
    I think that's only actually necessary if you want to
    find the two *shortest* disjoint paths, but all I care
    about is finding *any* two disjoint paths
3. Because of #2, I could replace dijkstra's algorithm
    (for finding the shortest path to *every* node) with
    the bidirectional variant (for only finding the path
    from the starting point to the end point)
4. Dropped the initial check for "is there a path from
    the car to the exit?" because the quick checker can
    already do that.
5. Experimented with only running Surballe's algorithm
    on every 10th step, with the thinking that it could
    still be useful, to detect and curtail routes that the
    quick check could not. However, running it often enough
    to be useful makes it slower than not running it.

I may have to look and see if I can find any quicker algorithms
for finding "any" path instead of "shortest" path. And then
I may be able to use the part of Surballe's that finds
disjointness, to check for two disjoint paths. I believe
the import thing about finding the disjoint paths, is just
that you have to reverse the direction of every edge in
the first path and give it weight 0. In other words, it'll
still work the same as the normal Surballe's, except with
an algorithm that finds any path instead of finding the
shortest.

But *is* it quicker to "find any path"?